account schema changes + bad request bug + cleanup

+++ src/main/java/com/vegapay/account/controller/AccountController.java
@@ -22,7 +22,7 @@ public class AccountController {
     }
 
     @GetMapping("/{accountId}")
-    public ResponseEntity<Account> getAccount(@PathVariable Long accountId) {
+    public ResponseEntity<Account> getAccount(@PathVariable int accountId) {
         Account account = accountService.getAccount(accountId);
         if (account != null) {
             return ResponseEntity.ok(account);
